requirements.txt
streamlit-0.80.0-py2.py3-none-any.whl
